+++ src/main/java/io/github/Gabriel/damagePlugin/customDamage/DamageListener.java
@@ -34,19 +34,18 @@ public class DamageListener implements Listener {
         ItemStack weapon = player.getInventory().getItemInMainHand();
         if (weapon == null || weapon.getType() == Material.AIR) return;
 
-        double totalDamage = 0.0;
-        Map<DamageType, Double> appliedTypes = new HashMap<>();
+        int totalDamage = 0;
+        Map<DamageType, Integer> appliedTypes = new HashMap<>();
 
         for (DamageType type : DamageType.values()) {
             if (damageKeys.checkForDamageKey(weapon, type)) {
-                double value = damageKeys.getDamageKeyValue(weapon, type);
+                int value = damageKeys.getDamageKeyValue(weapon, type);
                 totalDamage += value;
                 appliedTypes.put(type, value);
             }
         }
 
         if (appliedTypes.isEmpty()) {
-            // Fallback to vanilla damage if no elemental types
             return;
         }
 
@@ -56,7 +55,7 @@ public class DamageListener implements Listener {
         damagee.setMetadata("custom-damage-processing", new FixedMetadataValue(plugin, true));
 
         // Apply each elemental type separately (or summed, depending on how CustomDamager is set up)
-        for (Map.Entry<DamageType, Double> entry : appliedTypes.entrySet()) {
+        for (Map.Entry<DamageType, Integer> entry : appliedTypes.entrySet()) {
             customDamager.doCustomDamage(damagee, player, entry.getKey(), entry.getValue());
         }
